Split the source
